+++ decidim-core/app/views/decidim/application/_documents.html.erb
