+++ decidim-core/app/views/decidim/application/_photos.html.erb
