+++ decidim-core/app/views/decidim/participatory_processes/show.html.erb
@@ -10,11 +10,4 @@
     </div>
   </div>
 </div>
-<% if current_participatory_process.attachments.any? %>
-  <div class="row attachments">
-    <div class="columns large-8">
-      <%= render partial: "documents", locals: { documents: current_participatory_process.documents } %>
-      <%= render partial: "photos", locals: { photos: current_participatory_process.photos } %>
-    </div>
-  </div>
-<% end %>
+<%= render partial: "attachments", locals: { attachable: current_participatory_process } %>
