+++ decidim-core/config/locales/en.yml
@@ -27,6 +27,11 @@ en:
       user_groups:
         not_verified: Not verified
         verified: Verified
+    application:
+      documents:
+        related_documents: Related documents
+      photos:
+        related_photos: Related photos
     authorization_handlers:
       decidim/dummy_authorization_handler: Example authorization
       errors:
@@ -85,11 +90,6 @@ en:
     participatory_process_steps:
       index:
         process_steps: Process steps
-    participatory_processes:
-      documents:
-        related_documents: Related documents
-      photos:
-        related_photos: Related photos
     shared:
       login_modal:
         please_sign_in: Please sign in
