+++ decidim-core/lib/decidim/core.rb
@@ -13,6 +13,7 @@ module Decidim
   autoload :ResourceManifest, "decidim/resource_manifest"
   autoload :Resourceable, "decidim/resourceable"
   autoload :Authorable, "decidim/authorable"
+  autoload :Attachable, "decidim/attachable"
   autoload :Features, "decidim/features"
 
   include ActiveSupport::Configurable
