+++ decidim-meetings/app/models/decidim/meetings/meeting.rb
@@ -5,16 +5,13 @@ module Decidim
     # title, description and any other useful information to render a custom meeting.
     class Meeting < Meetings::ApplicationRecord
       include Decidim::Resourceable
+      include Decidim::Attachable
 
       belongs_to :feature, foreign_key: "decidim_feature_id", class_name: Decidim::Feature
       belongs_to :scope, foreign_key: "decidim_scope_id", class_name: Decidim::Scope
       belongs_to :category, foreign_key: "decidim_category_id", class_name: Decidim::Category
       has_one :organization, through: :feature
-      has_many :attachments,
-        class_name: Decidim::Attachment,
-        dependent: :destroy,
-        inverse_of: :attachable,
-        as: :attachable
+
       validates :title, presence: true
       validate :scope_belongs_to_organization
       validate :category_belongs_to_organization
