+++ decidim-meetings/app/views/decidim/meetings/meetings/show.html.erb
@@ -39,3 +39,4 @@
     <%= linked_resources_for meeting, :proposals, "proposals_from_meeting" %>
   </div>
 </div>
+<%= render partial: "attachments", locals: { attachable: meeting } %>
